+++ packages/polaris-test/nest-server-schema-first/graphql/resolvers/author.resolver.ts
@@ -54,10 +54,19 @@ export class AuthorResolver {
     }
 
     @Query()
-    public async onlinePaginatedAuthors(): Promise<OnlinePaginatedResolver<Author>> {
+    public async onlinePaginatedAuthorsWithLeftJoin(): Promise<OnlinePaginatedResolver<Author>> {
         return {
             getData: (): Promise<Author[]> => {
-                return this.authorService.findSortedByDataVersion();
+                return this.authorService.findWithLeftJoin();
+            },
+        };
+    }
+
+    @Query()
+    public async onlinePaginatedAuthorsWithInnerJoin(): Promise<OnlinePaginatedResolver<Author>> {
+        return {
+            getData: (): Promise<Author[]> => {
+                return this.authorService.findWithInnerJoin();
             },
         };
     }