+++ packages/polaris-test/shared-resources/entities/book.ts
@@ -2,14 +2,19 @@ import {
     Column,
     CommonModel,
     Entity,
+    JoinColumn,
+    ManyToMany,
     ManyToOne,
     OneToMany,
+    OneToOne,
     PrimaryGeneratedColumn,
     RelationId,
 } from '@enigmatis/polaris-core';
 import { Author } from './author';
 import { Chapter } from './chapter';
 import { Review } from './review';
+import { Genre } from './genre';
+import { OneToOneEntity } from './one-to-one-entity';
 
 @Entity()
 export class Book extends CommonModel {
@@ -27,6 +32,13 @@ export class Book extends CommonModel {
 
     @RelationId((book: Book) => book.chapters)
     public chaptersIds?: string[];
+
+    @ManyToMany((type) => Genre, (genre) => genre.books)
+    public genres: Genre[];
+
+    @OneToOne((type) => OneToOneEntity)
+    @JoinColumn()
+    public oneToOneEntity: OneToOneEntity;
 
     @PrimaryGeneratedColumn('uuid')
     protected id!: string;